Take care of last corner cases.
We were not searching for correct incoming/outgoing neighbors of reverse orientation of a segment. Now we do. Also, there were cases in one2one relation in the graph that we were keeping the middle segment while we could check a special case if we needed that middle segment at all or not.