Implementing the merging method instead of clipping off.
Not passing the validation tests yet. Some weird cases are happening .. self loops and stuff that needs a more focused mind to figure out. But has the main skeleton. merging criteria, handles different type of merging and corner cases, updated the graph so added a part that removes node from graph and updates all edges. This version is commited with some cerrs that I was outputing for debuging purposes. You can look if you're interested as they are not solved yet.