Improved heuristic for pos when not writing CIGARS
When we are not writing out the actual CIGAR string corresponding to the
optimal alignment, then it is better to write down the read position
assuming that the initial part of the read before the first MEM
is gapless.  Although this may not be what is encoded in the
actual, optimal alignemnt score, it will nonetheless ensure that
downstream tools place the rest of the matches in the read
appropriately.  One alternative is to modify the CIGAR string we write
down (even in the mode where we don't write optimal CIGAR strings),
to treat the intial read part as soft-clipped.  However, this may
be overly pessimistic if the optimal alignemnt of this piece doesn't
actually contain indels.  This commit is designed to address
https://github.com/COMBINE-lab/salmon/issues/475.