Add stricter checking of the decoys
This addresses https://github.com/mikelove/tximeta/issues/24
where an empty decoy file goes undetected because we previously
did not check for this.

Now, if the decoy file is empty, or if it contains decoy names
for which records are not found in the fasta file being
indexed, then fixFasta returns a non-zero value (and indexing halts).