Allow approxReadStartPos to return number < 0
If the read may overhang the transcript, then approxReadStarPos should
be allowed to return a number < 0.  This is important because this is
used in the dovetail check, and if we don't allow negative numbers here
then we may falsely prefer reads that _would_ be dovetail, but are not
marked as so because the reverse complement read overhangs the reference
while the forward read does not (or overhangs the transcript by more
nucleotides than does the forward read).

This fix simply removes the max() that is taken with 0, allowing this
function to return a negative number.  I checked that this function is
only called in two places.  In the first call, this should fix the
exisitng bug, and in the second (orphan recovery), the returned
"anchorPos" is already max()'d with 0 to compute the startPos that is
used to index into the transcript.  Therefore, the max() in this
function should not be necessary.